Combine VTK functions into `windtools.io.vtk`
